+++ CHANGELOG.md
@@ -83,6 +83,8 @@
   - fixed display of the description correctly
 - Add Roles
   - fixed empty chip display for additional roles in app overview active apps
+- App Roles
+  - update administration app roles api's
 - Company Certificate
   - Show inactive certificate information in details overlay
 - Fixed Automatic Up-Scrolling when Clicking on "More" Button in multiple pages